minor fixes

+++ src/bool_lexer.py
@@ -61,7 +61,7 @@ class bool_lexer(Lexer):
 	#@_(r'[\-]?\d+')
 	@_(r'[\-]?\d+([eE][-+]?\d+)?')
 	def INTEGER(self, t):
-		t.value = int(t.value)
+		t.value = float(t.value)
 		t.type = INTEGER
 		return t
